fix PR review Todo fix swarm server
