+++ common/config/config.json
@@ -3,5 +3,7 @@
     "latest": "0.13.1",
     "stable": "0.13.0"
   },
-  "storage": "ipfs"
+  "storage": "ipfs",
+  "ipfsGateway": "https://ipfs.io/ipfs/",
+  "swarmGateway": "https://api.gateway.ethswarm.org/bzz/"
 }
