+++ common/lib/kbeejs.ts
@@ -1,17 +1,20 @@
 import { Bee, Data, FileData } from "@ethersphere/bee-js";
-import { swarmLinkToCid, SWARM_GATEWAY, SWARM_GATEWAY_BZZ } from "./kconfig";
+import { swarmLinkToCid, swarmServer, SWARM_GATEWAY } from "./kconfig";
 
+const getBee = (nodeUrl: string): Bee => {
+  return new Bee(nodeUrl ? nodeUrl : swarmServer(SWARM_GATEWAY));
+};
 const krdbatchId = "0000000000000000000000000000000000000000000000000000000000000000";
 
 const swarmUploadFile = async (
   file: File | string,
   fileName: string,
   contentType: string,
-  nodeUrl?: string,
+  nodeUrl = "",
   batchId?: string,
   fileSize?: number
 ): Promise<string> => {
-  const bee: Bee = new Bee(nodeUrl ? nodeUrl : SWARM_GATEWAY);
+  const bee: Bee = getBee(nodeUrl);
 
   const result = await bee.uploadFile(batchId ? batchId : krdbatchId, file, fileName, {
     pin: nodeUrl ? true : false,
@@ -22,17 +25,17 @@ const swarmUploadFile = async (
   return result.reference;
 };
 
-const swarmDownloadFile = async (fileReference: string, nodeUrl?: string): Promise<FileData<Data>> => {
-  const bee: Bee = new Bee(nodeUrl ? nodeUrl : SWARM_GATEWAY);
+const swarmDownloadFile = async (fileReference: string, nodeUrl = ""): Promise<FileData<Data>> => {
+  const bee: Bee = getBee(nodeUrl);
 
   return await bee.downloadFile(fileReference);
 };
 
-const swarmGetContentType = async (fileReference: string, nodeUrl?: string): Promise<string> => {
-  const bee: Bee = new Bee(nodeUrl ? nodeUrl : SWARM_GATEWAY);
+const swarmGetContentType = async (fileReference: string, nodeUrl = ""): Promise<string> => {
+  const bee: Bee = getBee(nodeUrl);
 
-  if (fileReference.startsWith(SWARM_GATEWAY_BZZ)) {
-    fileReference = fileReference.replace(SWARM_GATEWAY_BZZ, "");
+  if (fileReference.startsWith(SWARM_GATEWAY)) {
+    fileReference = fileReference.replace(SWARM_GATEWAY, "");
   } else if (fileReference.startsWith("swarm://")) {
     fileReference = swarmLinkToCid(fileReference);
   }
